+++ src/pipx/commands/run.py
@@ -19,7 +19,7 @@ from pipx.util import (
     rmdir,
     run_pypackage_bin,
 )
-from pipx.venv import PackageInstallFailureError, Venv
+from pipx.venv import Venv
 
 
 def run(
@@ -119,17 +119,14 @@ def _download_and_run(
     #   pre-existing, otherwise is None to instruct venv.install_package to
     #   determine package name.
 
-    try:
-        venv.install_package(
-            package=venv.pipx_metadata.main_package.package,
-            package_or_url=package_or_url,
-            pip_args=pip_args,
-            include_dependencies=False,
-            include_apps=True,
-            is_main_package=True,
-        )
-    except PackageInstallFailureError:
-        raise PipxError(f"Unable to install {package_or_url}")
+    venv.install_package(
+        package=venv.pipx_metadata.main_package.package,
+        package_or_url=package_or_url,
+        pip_args=pip_args,
+        include_dependencies=False,
+        include_apps=True,
+        is_main_package=True,
+    )
 
     if not (venv.bin_path / app).exists():
         apps = venv.pipx_metadata.main_package.apps